Refactor interp utils to prediction.py
